Refactor: separation of utility functions
getContrast 함수 재사용하기 위해 분리
