+++ frontend/src/component/common/Label.js
@@ -1,32 +1,15 @@
 import React from 'react'
 import styled from 'styled-components'
+import { getContrast } from '@Util/util.js'
 const Label = ({ name, color }) => (
   <StyledLabel backgroundColor={color}>{name}</StyledLabel>
 )
-
-const getContrast = hexColor => {
-  if (hexColor.startsWith('#')) hexColor = hexColor.slice(1)
-  if (hexColor.length === 3)
-    hexColor = hexColor
-      .split('')
-      .map(hex => hex + hex)
-      .join('')
-  return getContrastFromRGB(...hexToRGB(hexColor)) >= 128 ? 'black' : 'white'
-}
-
-const getContrastFromRGB = (r, g, b) => (r * 299 + g * 587 + b * 114) / 1000
-
-const hexToRGB = hexColor => [
-  parseInt(hexColor.slice(0, 2), 16),
-  parseInt(hexColor.slice(2, 4), 16),
-  parseInt(hexColor.slice(4, 6), 16),
-]
 
 const StyledLabel = styled.div`
   ${({ backgroundColor }) =>
     `
     background-color: ${backgroundColor};
-    color: ${getContrast(backgroundColor.slice(1))};
+    color: ${getContrast(backgroundColor)};
   `}
   padding: 0 7px;
   font-size: 12px;
