+++ frontend/src/util/util.js
@@ -28,4 +28,22 @@ const getMilestoneDateFormat = date => {
   return `Due by ${monthString[parseInt(month)]} ${day}, ${year}`
 }
 
-export { getTimePassedFromNow, getMilestoneDateFormat }
+const getContrast = hexColor => {
+  if (hexColor.startsWith('#')) hexColor = hexColor.slice(1)
+  if (hexColor.length === 3)
+    hexColor = hexColor
+      .split('')
+      .map(hex => hex + hex)
+      .join('')
+  return getContrastFromRGB(...hexToRGB(hexColor)) >= 128 ? 'black' : 'white'
+}
+
+const getContrastFromRGB = (r, g, b) => (r * 299 + g * 587 + b * 114) / 1000
+
+const hexToRGB = hexColor => [
+  parseInt(hexColor.slice(0, 2), 16),
+  parseInt(hexColor.slice(2, 4), 16),
+  parseInt(hexColor.slice(4, 6), 16),
+]
+
+export { getTimePassedFromNow, getMilestoneDateFormat, getContrast }
